FEAT: 코드 추가
퀘스트 로직 추가, 구조변경
